[#47014] Replace Settings::XxxSettingCell by ViewComponent
https://community.openproject.org/work_packages/47014
