--- app/cells/components/on_off_status_cell.rb
@@ -1,16 +1,0 @@
-module Components
-  ##
-  # A language switch and text area for updating a localized text setting.
-  class OnOffStatusCell < ::RailsCell
-    options :on_text
-    options :on_description
-    options :off_text
-    options :off_description
-
-    options :is_on
-
-    def enabled?
-      !!model[:is_on]
-    end
-  end
-end
