--- app/cells/views/settings/numeric_setting/show.erb
@@ -1,11 +1,0 @@
-<div class="form--field">
-  <%=
-    setting_text_field(
-      name,
-      size: size,
-      unit: t("label_#{unit.singularize}_plural"),
-      container_class: '-xslim',
-      title: title
-    )
-  %>
-</div>
