--- app/cells/views/settings/time_zone_setting/show.erb
@@ -1,3 +1,0 @@
-<div class="form--field">
-  <%= render_select %>
-</div>
