+++ app/components/components/on_off_status_component.html.erb
@@ -26,19 +26,23 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
-<%= fields_for :pref, @user.pref, builder: TabularFormBuilder, lang: current_language do |pref_fields| %>
-  <%= render Settings::TimeZoneSettingComponent.new(
-        "time_zone",
-        form: pref_fields,
-        container_class: (defined? input_size) ? "-#{input_size}" : "-wide"
-      )
-  %>
-  <div class="form--field">
-    <%= pref_fields.select :comments_sorting,
-                           [[t(:label_chronological_order), 'asc'],
-                            [t(:label_reverse_chronological_order), 'desc']],
-                           container_class: (defined? input_size) ? "-#{input_size}" : '' %>
-  </div>
-  <div class="form--field"><%= pref_fields.check_box :warn_on_leaving_unsaved %></div>
-  <div class="form--field"><%= pref_fields.check_box :auto_hide_popups %></div>
-<% end %>
+
+<%# The javscript for this cell is included in `settings.js.erb`. %>
+
+<div class="on-off-status-cell">
+  <% if enabled? %>
+  <span class="on-off-status -enabled">
+    <%= op_icon 'icon-yes' %>
+    <%= model[:on_text] %>
+  </span>
+      <p>
+        <%= model[:on_description] %>
+      </p>
+  <% else %>
+  <span class="on-off-status -disabled">
+    <%= op_icon 'icon-not-supported' %>
+    <%= model[:off_text] %>
+  </span>
+      <p><%= model[:off_description] %></p>
+  <% end %>
+</div>
