+++ app/components/settings/time_zone_setting_component.html.erb
@@ -26,19 +26,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
-<%= fields_for :pref, @user.pref, builder: TabularFormBuilder, lang: current_language do |pref_fields| %>
-  <%= render Settings::TimeZoneSettingComponent.new(
-        "time_zone",
-        form: pref_fields,
-        container_class: (defined? input_size) ? "-#{input_size}" : "-wide"
-      )
-  %>
-  <div class="form--field">
-    <%= pref_fields.select :comments_sorting,
-                           [[t(:label_chronological_order), 'asc'],
-                            [t(:label_reverse_chronological_order), 'desc']],
-                           container_class: (defined? input_size) ? "-#{input_size}" : '' %>
-  </div>
-  <div class="form--field"><%= pref_fields.check_box :warn_on_leaving_unsaved %></div>
-  <div class="form--field"><%= pref_fields.check_box :auto_hide_popups %></div>
-<% end %>
+
+<div class="form--field">
+  <%= render_select %>
+</div>
