+++ app/views/admin/settings/authentication_settings/show.html.erb
@@ -49,12 +49,12 @@ See COPYRIGHT and LICENSE files for more details.
     </fieldset>
 
     <% if OpenProject::Configuration.registration_footer.blank? %>
-      <%= rails_cell Settings::NumericSettingCell, "invitation_expiration_days", unit: "days" %>
+      <%= render Settings::NumericSettingComponent.new("invitation_expiration_days", unit: "days") %>
 
       <fieldset class="form--fieldset">
         <fieldset id="registration_footer" class="form--fieldset">
           <legend class="form--fieldset-legend"><%= I18n.t(:setting_registration_footer) %></legend>
-          <%= rails_cell Settings::TextSettingCell, I18n.locale, name: "registration_footer" %>
+          <%= render Settings::TextSettingComponent.new(I18n.locale, name: "registration_footer") %>
         </fieldset>
       </fieldset>
     <% end %>
