+++ app/views/admin/settings/mail_notifications_settings/show.html.erb
@@ -40,8 +40,8 @@ See COPYRIGHT and LICENSE files for more details.
     </section>
 
     <fieldset id="emails_decorators" class="form--fieldset"><legend class="form--fieldset-legend"><%= t(:setting_emails_header) %> & <%= t(:setting_emails_footer) %></legend>
-      <%= rails_cell Settings::TextSettingCell, I18n.locale, name: "emails_header" %>
-      <%= rails_cell Settings::TextSettingCell, I18n.locale, name: "emails_footer" %>
+      <%= render Settings::TextSettingComponent.new(I18n.locale, name: "emails_header") %>
+      <%= render Settings::TextSettingComponent.new(I18n.locale, name: "emails_footer") %>
     </fieldset>
   <% else %>
     <div class="nodata">
