+++ app/views/admin/settings/users_settings/show.html.erb
@@ -39,13 +39,11 @@ See COPYRIGHT and LICENSE files for more details.
       <%= setting_select :default_language, all_lang_options_for_select(false), container_class: '-slim' %>
     </div>
 
-    <%=
-      cell(
-          Settings::TimeZoneSettingCell,
+    <%= render Settings::TimeZoneSettingComponent.new(
           "user_default_timezone",
           container_class: "-slim",
           title: I18n.t("tooltip_user_default_timezone")
-      )
+        )
     %>
 
     <div class="form--field">
@@ -71,7 +69,7 @@ See COPYRIGHT and LICENSE files for more details.
       <%= setting_check_box :consent_required %>
     </div>
 
-    <%= rails_cell Settings::TextSettingCell, I18n.locale, name: "consent_info" %>
+    <%= render Settings::TextSettingComponent.new(I18n.locale, name: "consent_info") %>
 
     <div class="form--field">
       <%= setting_block("consent_time") do %>
