+++ app/views/users/_consent.html.erb
@@ -1,11 +1,10 @@
 <section class="form--section">
   <h3 class="form--section-title"><%=t('consent.title')%></h3>
-  <%= rails_cell ::Components::OnOffStatusCell,
-                 {
-                   is_on: @user.consented_at.present?,
-                   on_text: format_time(@user.consented_at),
-                   on_description: t('consent.user_has_consented'),
-                   off_text: t(:label_never),
-                   off_description: t('consent.not_yet_consented')
-                 } %>
+  <%= render ::Components::OnOffStatusComponent.new({
+               is_on: @user.consented_at.present?,
+               on_text: format_time(@user.consented_at),
+               on_description: t('consent.user_has_consented'),
+               off_text: t(:label_never),
+               off_description: t('consent.not_yet_consented')
+             }) %>
 </section>
