+++ modules/two_factor_authentication/app/views/two_factor_authentication/my/two_factor_devices/_two_factor_status.html.erb
@@ -1,8 +1,7 @@
-<%= rails_cell ::Components::OnOffStatusCell,
-               { is_on: device.present?,
-                 on_text: t('two_factor_authentication.label_2fa_enabled'),
-                 on_description: t('two_factor_authentication.text_2fa_enabled'),
-                 off_text: t('two_factor_authentication.label_2fa_disabled'),
-                 off_description: t('two_factor_authentication.text_2fa_disabled')
-               }%>
-
+<%= render ::Components::OnOffStatusComponent.new({
+             is_on: device.present?,
+             on_text: t('two_factor_authentication.label_2fa_enabled'),
+             on_description: t('two_factor_authentication.text_2fa_enabled'),
+             off_text: t('two_factor_authentication.label_2fa_disabled'),
+             off_description: t('two_factor_authentication.text_2fa_disabled')
+           })%>
